requirements.txt
fastapi
uvicorn
openai==0.28.1
python-dotenv
pytest
httpx
pytest-cov
motor
